server/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
alembic>=1.13.0
pydantic-settings>=2.0.0
cryptography>=42.0.0
jinja2>=3.1.0
